Add experimental wasm bindings
Published to npm as objdiff-wasm

+++ objdiff-core/src/arch/arm.rs
@@ -111,7 +111,7 @@ impl ObjArch for ObjArchArm {
         code: &[u8],
         section_index: usize,
         relocations: &[ObjReloc],
-        line_info: &BTreeMap<u64, u64>,
+        line_info: &BTreeMap<u64, u32>,
         config: &DiffObjConfig,
     ) -> Result<ProcessCodeResult> {
         let start_addr = address as u32;
