+++ objdiff-core/src/arch/ppc.rs
@@ -35,7 +35,7 @@ impl ObjArch for ObjArchPpc {
         code: &[u8],
         _section_index: usize,
         relocations: &[ObjReloc],
-        line_info: &BTreeMap<u64, u64>,
+        line_info: &BTreeMap<u64, u32>,
         config: &DiffObjConfig,
     ) -> Result<ProcessCodeResult> {
         let ins_count = code.len() / 4;
